requirements.txt
colorama==0.4.6
openai==1.30.4
python-dotenv==1.0.1
